+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.3
+
+### Sunday, 6th December, 2022
+
+* Fix: Unable to load asset.
+* Fix: Active Step radius.
+
 ## 0.1.2
 
 ### Sunday, 5th December, 2022